requirements.txt
Flask
gunicorn
pyyaml==6.0.1
jinja2==3.1.2
slack-sdk==3.23.0
prometheus-client==0.18.0
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
docker==6.1.3
